pairの型チェックがVS2017で厳しくなった(C4244)ので対応

+++ BonCtrl/PMTUtil.cpp
@@ -96,7 +96,7 @@ BOOL CPMTUtil::DecodePMT(BYTE* data, DWORD dataSize)
 				//CA
 				WORD CA_PID = ((WORD)data[readSize+2]&0x1F)<<8 | (WORD)data[readSize+3];
 				if (CA_PID != 0x1fff) {
-					PIDList.insert(pair<WORD,WORD>(CA_PID, 0));
+					PIDList.insert(pair<WORD,WORD>(CA_PID, (WORD)0));
 				}
 			}
 			readSize += descriptor_length;
@@ -123,7 +123,7 @@ BOOL CPMTUtil::DecodePMT(BYTE* data, DWORD dataSize)
 					//CA
 					WORD CA_PID = ((WORD)data[readSize+2]&0x1F)<<8 | (WORD)data[readSize+3];
 					if (CA_PID != 0x1fff) {
-						PIDList.insert(pair<WORD,WORD>(CA_PID, 0));
+						PIDList.insert(pair<WORD,WORD>(CA_PID, (WORD)0));
 					}
 				}
 				readSize += descriptor_length;
